powershell: add jj completion as a background job
Executing jj twice to add static and dynamic completions adds several
hundred milliseconds to shell startup.

It's very unlikely that completion will be needed immediately, so
instead execute jj in background jobs to generate the scripts that add
completion. Then, the Prompt function checks if the jobs have finished
and invokes the script it returns.